fix(nx-ignore): check that yarn/pnpm are installed before using on Netlify (#362)

+++ packages/nx-ignore/src/index.ts
@@ -166,11 +166,20 @@ function installTempNx(root: string, plugins: string[]): string | null {
     }
     writeFileSync(join(tmpPath, 'package.json'), JSON.stringify(json));
 
-    if (existsSync(join(root, 'yarn.lock'))) {
+    if (
+      existsSync(join(root, 'yarn.lock')) &&
+      isPackageManagerInstalled(`yarn`)
+    ) {
+      logDebug(`Using yarn to install Nx.`);
       execSync(`yarn install`, { cwd: tmpPath });
-    } else if (existsSync(join(root, 'pnpm-lock.yaml'))) {
+    } else if (
+      existsSync(join(root, 'pnpm-lock.yaml')) &&
+      isPackageManagerInstalled(`pnpm`)
+    ) {
+      logDebug(`Using pnpm to install Nx.`);
       execSync(`pnpm install --force`, { cwd: tmpPath });
     } else {
+      logDebug(`Using npm to install Nx.`);
       execSync(`npm install --force`, { cwd: tmpPath });
     }
     moveSync(join(tmpPath, 'node_modules'), join(root, 'node_modules'));
@@ -181,6 +190,19 @@ function installTempNx(root: string, plugins: string[]): string | null {
   }
 
   return null;
+}
+
+function isPackageManagerInstalled(pm: 'yarn' | 'pnpm'): boolean {
+  try {
+    const version = execSync(`${pm} --version`, {
+      stdio: 'pipe',
+    })?.toString();
+    logDebug(`Found ${pm} version ${version}.`);
+    return true;
+  } catch {
+    logDebug(`Could not find ${pm}. Check that it is installed.`);
+    return false;
+  }
 }
 
 function commitHasSkipMessage(message: string): boolean {
